Move dict_almost_equal to io.py due to not being able to import it in /tests/tracing/*.py
